add prefix filter to cleanup

+++ tests/cleanup_resources.py
@@ -11,16 +11,18 @@ if __name__ == "__main__":
     buckets = client.list_buckets()
 
     for bucket in buckets["Buckets"]:
-        s3_bucket = s3.Bucket(bucket["Name"])
-        s3_bucket.objects.all().delete()
-        s3_bucket.object_versions.delete()
-        s3_bucket.delete()
-        print(f"Deleted s3 bucket {bucket['Name']}")
+        if bucket["Name"].startswith("pytest"):
+            s3_bucket = s3.Bucket(bucket["Name"])
+            s3_bucket.objects.all().delete()
+            s3_bucket.object_versions.delete()
+            s3_bucket.delete()
+            print(f"Deleted s3 bucket {bucket['Name']}")
 
     storage_client = storage.Client()
     buckets = storage_client.list_buckets()
     for bucket in buckets:
-        bucket.delete(force=True)
-        print(f"Deleted gcp bucket {bucket.name}")
+        if bucket.name.startswith("pytest"):
+            bucket.delete(force=True)
+            print(f"Deleted gcp bucket {bucket.name}")
 
     print(f"Not deleting Azure containers")
